MODIFY:
-> Moving them to "Methods" folder
-> ADD the 'deactivated_users_list' in deactivate_user.py file
